Fix for: Unable to import root categories.
The Import Feature Says that a Category Must have a Parent Category (ParentCategoryId is Required).
But a Root level category does not have a ParentCategory.

This throws a Missing Required Property error when you try to import a root level category.

I have made a change in CategoryImporter.cs. Changed categoryParentId IsRequiredProperty from
true to false.