VP-3719: Allow platofrm to start when cannot copy same version assembly (#1948)
In some cases (e.g. multiple instances sharing one probing folder) the platform could be unable to start because each instance is trying to write its version of the file. Need to allow the platform to start in case of versions are the same, only modification dates differ. Still, warnings about such facts would be logged.